fix(rendering): silence optional terrain runtime context
